Split into Client and Server threads
This commit introduces a Client-Server architecture to the game server,
enabling concurrency via the use of threads.

The code is split into modules for better seperation of concern,
improved modularity, and easier code navigation.

A basic Game Protocol is defined, with seperate Client and Server
events. This seperation allows for specific pattern matching on
the receiving side.

The distinct players of a game are now identified by numeric IDs, which are
passed in messages. On the client side, each player is associated with
an 'icon' to display the player on the game board. On the server
side, each player is associated with a channel Sender and Receiver,
enabling communication with each player individually. This design allows
for multiple client connections to the server, one for each player.

The state of the game board is now calculated on-demand instead of being
stored. This reduces the number of required mutable references to the board, as storing
provides minimal benefits.