Fix flaky unit test
Due to the unpredictability of asynchronous code, certain unit tests were flaky as they were attempting to guess the order that messages would be received into a Stream.

This caused test failures in the case that messages were received in the opposite order to what the test expected.

To fix this, we generalise the message assertions by conditionally asserting the expected outcome based on the given player id.